requirements.txt
streamlit==1.24.0
Pillow
numpy
keras
tensorflow
matplotlib